requirements.txt
aiogram==2.22.1
requests==2.28.1
beautifulsoup4==4.11.1
pyppeteer==1.0.2
Whoosh==2.7.4
boto3==1.24.69
Pyrogram==2.0.51
Pillow==9.2.0
tgcrypto==1.2.3
Jinja2==3.1.2
python-dotenv==0.21.0